requirements.txt
chromadb
google-generativeai
pypdf
pillow
python-dotenv
tqdm
transformers

torch
torchvision

git-filter-repo